Use Yarn package manager (#47282)
* Use Yarn package manager

* Fix travis build

* Remove Babel macros, support for Node 12.13.0 LTS
