+++ tgui-next/README.md
@@ -36,7 +36,8 @@ highlights differences with React.
 
 You will need these programs to start developing in tgui:
 
-- [Node 12.x](https://nodejs.org/en/download/current/)
+- [Node v12.10+](https://nodejs.org/en/download/current/)
+- [Yarn v1.19+](https://yarnpkg.com/en/docs/install)
 - [MSys2](https://www.msys2.org/) (optional)
 
 > MSys2 closely replicates a unix-like environment which is necessary for
@@ -52,14 +53,14 @@ If you haven't opened the console already, you can do that by holding
 Shift and right clicking on the `tgui-next` folder, then pressing
 either `Open command window here` or `Open PowerShell window here`.
 
-Run `npm install`, then:
+Run `yarn install`, then:
 
-- `npm run build` - build the project in production mode.
-- `npm run watch` - launch a development server, with live log collection,
+- `yarn run build` - build the project in production mode.
+- `yarn run watch` - launch a development server, with live log collection,
 cache reloading and hot module replacement.
-- `npm run lint` - show problems with the code.
-- `npm run lint --fix` - auto-fix problems with the code.
-- `npm run analyze` - run a bundle analyzer.
+- `yarn run lint` - show problems with the code.
+- `yarn run lint --fix` - auto-fix problems with the code.
+- `yarn run analyze` - run a bundle analyzer.
 
 For MSys2, WSL, Linux or macOS users:
 
@@ -75,10 +76,11 @@ cache reloading and hot module replacement.
 - `bin/tgui [webpack options]` - build the project with custom webpack
 options.
 
-For absolute brainlets, we also got a batch file in store. Double click
-it to build the project:
+We also got some batch files in store, which are simple to use:
 
 - `bin/tgui-build.bat` - build the project in production mode.
+- `bin/tgui-watch.bat` - launch a development server, with live log
+collection, cache reloading and hot module replacement.
 
 Remember to always run a full build before submitting a PR. It creates
 a compressed javascript bundle which is then referenced from DM code.
