+++ tgui-next/bin/tgui-watch.bat
@@ -1,5 +1,4 @@
 @echo off
 cd "%~dp0\.."
 call yarn install
-call yarn run build
-timeout /t 9
+call yarn run watch
