--- tgui-next/lerna.json
@@ -1,6 +1,0 @@
-{
-  "packages": [
-    "packages/*"
-  ],
-  "version": "0.1.0"
-}
