+++ tgui-next/packages/tgui/package.json
@@ -8,7 +8,6 @@
     "@babel/preset-env": "^7.6.2",
     "babel-loader": "^8.0.6",
     "babel-plugin-inferno": "^6.1.0",
-    "babel-plugin-macros": "^2.6.1",
     "core-js": "^3.2.1",
     "css-loader": "^3.2.0",
     "cssnano": "^4.1.10",
