+++ tools/travis/build_tgui.sh
@@ -5,8 +5,14 @@ set -euo pipefail
 cd "$(dirname "${0}")/../.."
 base_dir="$(pwd)"
 
+## The final authority on what's required to fully build the project
+source dependencies.sh
+
 ## Setup NVM
-[[ -e ~/.nvm/nvm.sh ]] && source ~/.nvm/nvm.sh
+if [[ -e ~/.nvm/nvm.sh ]]; then
+	source ~/.nvm/nvm.sh
+	nvm use "${NODE_VERSION}"
+fi
 
 echo "Building 'tgui'"
 cd "${base_dir}/tgui"
