+++ tools/travis/install_build_tools.sh
@@ -5,6 +5,8 @@ source dependencies.sh
 
 source ~/.nvm/nvm.sh
 nvm install $NODE_VERSION
+nvm use $NODE_VERSION
+npm install --global yarn
 
 pip3 install --user PyYaml
 pip3 install --user beautifulsoup4
